Improve world info scraper
