+++ .gitignore
@@ -1,12 +1,7 @@
-Racing Club/__pycache__/
-Racing Club/report/
-Racing Club/data/
+track_results/__pycache__/
+track_results/report/
+track_results/data/
 
-VRChat World Info/scraper/*.json
-VRChat World Info/__pycache__/
-VRChat World Info/docs/approved_export.json
-
-data/
-report/
-site/
-__pycache__/
+world_info/scraper/*.json
+world_info/__pycache__/
+world_info/docs/approved_export.json
