+++ README.md
@@ -2,18 +2,18 @@
 
 This repository contains two separate toolsets:
 
-## Racing Club
+## Track Results
 Utilities that download racing records from a Google Spreadsheet and
-build text or HTML leaderboards. See [`Racing Club/README.md`](Racing%20Club/README.md)
+build text or HTML leaderboards. See [`track_results/README.md`](track_results/README.md)
 for full usage instructions.
 
-## VRChat World Info
+## World Info
 A workflow for collecting VRChat world information and approving entries.
 It exports a JSON file for use on a website or inside Unity and includes a
 simple Tkinter interface for reviewing worlds.
-See [`VRChat World Info/README.md`](VRChat%20World%20Info/README.md) for details.
+See [`world_info/README.md`](world_info/README.md) for details.
 Additional background is available in
-[`VRChat World Info/complete_guide.zh_TW.md`](VRChat%20World%20Info/complete_guide.zh_TW.md) (Chinese).
+[`world_info/complete_guide.zh_TW.md`](world_info/complete_guide.zh_TW.md) (Chinese).
 
 For Traditional Chinese instructions, read
 [`README.zh_TW.md`](README.zh_TW.md).
