--- Racing Club/site/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>VR Racing Club Leaderboard</title>
-    <style>
-        body {font-family: sans-serif; padding: 20px;}
-        pre {background: #f4f4f4; padding: 10px; white-space: pre-wrap;}
-    </style>
-</head>
-<body>
-<h1>VR Racing Club Leaderboard</h1>
-<pre>No data
-</pre>
-</body>
-</html>
