+++ world_info/README.md
@@ -4,7 +4,7 @@ This tool collects information about VRChat worlds, lets you review the entries
 and exports a filtered JSON file for use on a website or in Unity.
 
 ```
-VRChat World Info/
+world_info/
 ├─ scraper/
 │  ├─ scraper.py          # query the VRChat API for world info
 │  ├─ review_tool.py      # mark worlds as approved
@@ -18,12 +18,17 @@ VRChat World Info/
    └─ GenerateWorldCards.cs
 ```
 
-Run the scripts in order:
+Create ``scraper/headers.json`` with your login cookie::
 
-1. `python3 scraper/scraper.py --keyword Taiwan` to search worlds by keyword
-   or `python3 scraper/scraper.py --user usr_abc123` to fetch worlds from a creator.
-2. `python3 scraper/review_tool.py` (or run `python3 ui.py` for a GUI)
-3. `python3 scraper/exporter.py`
+  {"Cookie": "auth=...; twoFactorAuth=...; machineId=..."}
+
+Run the tools in order:
+
+1. ``python3 scraper/scraper.py --keyword Taiwan --limit 50`` to search worlds
+   or ``python3 scraper/scraper.py --user usr_abc123 --limit 50`` to fetch a
+   creator's worlds.  Results are written to ``raw_worlds.json``.
+2. ``python3 scraper/review_tool.py`` (or run ``python3 ui.py`` for a GUI)
+3. ``python3 scraper/exporter.py``
 
 Copy `scraper/approved_export.json` into `docs/` to update the website or load
 it inside Unity using the `GenerateWorldCards` editor script.
