+++ world_info/README.zh_TW.md
@@ -4,7 +4,7 @@
 並匯出篩選後的 JSON 檔，可用於網站或 Unity 專案。
 
 ```
-VRChat World Info/
+world_info/
 ├─ scraper/
 │  ├─ scraper.py          # 從 VRChat API 取得世界資料
 │  ├─ review_tool.py      # 標記世界是否核可
@@ -18,10 +18,17 @@ VRChat World Info/
    └─ GenerateWorldCards.cs
 ```
 
-執行順序：
+請在 ``scraper/headers.json`` 中輸入登入後取得的 Cookie，例如：
 
-1. `python3 scraper/scraper.py --keyword Taiwan` 以關鍵字搜尋世界，
-   或 `python3 scraper/scraper.py --user usr_abc123` 取得指定作者世界。
+```
+{"Cookie": "auth=...; twoFactorAuth=...; machineId=..."}
+```
+
+執行流程：
+
+1. `python3 scraper/scraper.py --keyword Taiwan --limit 50` 以關鍵字搜尋世界，
+   或 `python3 scraper/scraper.py --user usr_abc123 --limit 50` 取得指定作者世界，
+   結果會輸出到 `raw_worlds.json`。
 2. `python3 scraper/review_tool.py`（或執行 `python3 ui.py` 使用圖形介面）
 3. `python3 scraper/exporter.py`
